+++ src/Element/Conditions/NotRelatedToConditionRule.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace craft\elements\conditions;
+namespace CraftCms\Cms\Element\Conditions;
 
 use craft\base\ElementInterface;
 use CraftCms\Cms\Element\Queries\Contracts\ElementQueryInterface;
